+++ cmd/zetaclientd/init.go
@@ -1,12 +1,24 @@
 package main
 
 import (
+	"encoding/json"
+	"os"
+	"path"
+
 	"github.com/rs/zerolog"
 	"github.com/spf13/cobra"
 
 	"github.com/zeta-chain/zetacore/zetaclient/config"
 	"github.com/zeta-chain/zetacore/zetaclient/testutils"
 )
+
+// solanaTestKey is a local test private key for Solana
+var solanaTestKey = []uint8{
+	199, 16, 63, 28, 125, 103, 131, 13, 6, 94, 68, 109, 13, 68, 132, 17,
+	71, 33, 216, 51, 49, 103, 146, 241, 245, 162, 90, 228, 71, 177, 32, 199,
+	31, 128, 124, 2, 23, 207, 48, 93, 141, 113, 91, 29, 196, 95, 24, 137,
+	170, 194, 90, 4, 124, 113, 12, 222, 166, 209, 119, 19, 78, 20, 99, 5,
+}
 
 var InitCmd = &cobra.Command{
 	Use:   "init",
@@ -36,6 +48,7 @@ type initArguments struct {
 	KeyringBackend      string
 	HsmMode             bool
 	HsmHotKey           string
+	SolanaKey           string
 }
 
 func init() {
@@ -69,6 +82,7 @@ func init() {
 	InitCmd.Flags().BoolVar(&initArgs.HsmMode, "hsm-mode", false, "enable hsm signer, default disabled")
 	InitCmd.Flags().
 		StringVar(&initArgs.HsmHotKey, "hsm-hotkey", "hsm-hotkey", "name of hotkey associated with hardware security module")
+	InitCmd.Flags().StringVar(&initArgs.SolanaKey, "solana-key", "solana-key.json", "solana key file name")
 }
 
 func Initialize(_ *cobra.Command, _ []string) error {
@@ -106,8 +120,36 @@ func Initialize(_ *cobra.Command, _ []string) error {
 	configData.KeyringBackend = config.KeyringBackend(initArgs.KeyringBackend)
 	configData.HsmMode = initArgs.HsmMode
 	configData.HsmHotKey = initArgs.HsmHotKey
+	configData.SolanaKeyFile = initArgs.SolanaKey
 	configData.ComplianceConfig = testutils.ComplianceConfigTest()
 
-	//Save config file
+	// Save solana fee payer key
+	keyFile := path.Join(rootArgs.zetaCoreHome, initArgs.SolanaKey)
+	err = createSolanaKeyFile(keyFile)
+	if err != nil {
+		return err
+	}
+
+	// Save config file
 	return config.Save(&configData, rootArgs.zetaCoreHome)
+}
+
+// createSolanaKeyFile creates a solana key json file
+func createSolanaKeyFile(keyFile string) error {
+	// marshal the byte array to JSON
+	keyBytes, err := json.Marshal(solanaTestKey)
+	if err != nil {
+		return err
+	}
+
+	// create file (or overwrite if it already exists)
+	file, err := os.Create(keyFile)
+	if err != nil {
+		return err
+	}
+	defer file.Close()
+
+	// write the key bytes to the file
+	_, err = file.Write(keyBytes)
+	return err
 }